+++ src/models/other_licensing_information_detected.rs
@@ -5,7 +5,7 @@
 use serde::{Deserialize, Serialize};
 
 /// <https://spdx.github.io/spdx-spec/6-other-licensing-information-detected/>
-#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+#[derive(Debug, Serialize, Deserialize, Clone, Default, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OtherLicensingInformationDetected {
     /// <https://spdx.github.io/spdx-spec/6-other-licensing-information-detected/#61-license-identifier>